+++ src/main/java/io/github/tufkan1/projectex/ProjectEX.java
@@ -5,6 +5,7 @@ import io.github.tufkan1.projectex.command.EmcCommands;
 import io.github.tufkan1.projectex.emc.mapping.minecraft.MinecraftRecipeMappingService;
 import io.github.tufkan1.projectex.emc.reload.EmcDataReloadListener;
 import io.github.tufkan1.projectex.internal.emc.EmcValueRegistry;
+import io.github.tufkan1.projectex.network.AlchemyNetworking;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.resource.v1.ResourceLoader;
 import net.minecraft.resources.Identifier;
@@ -26,6 +27,7 @@ public final class ProjectEX implements ModInitializer {
             new EmcDataReloadListener(EMC_VALUES)
         );
         EmcCommands.register();
+        AlchemyNetworking.register();
         MinecraftRecipeMappingService.register(EMC_VALUES);
         LOGGER.info("ProjectEX {} is initializing with {} EMC values", version(), EMC_VALUES.snapshot().size());
     }